# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Stylist — Stylist Backend
  version: 0.3.3
  summary: Backend that holds the OpenRouter key and runs the deterministic recommendation pipeline. Keeps no user data; its only state is a cost ledger and hashed device-token registry.
  description: |
    The contract between the iOS client and the stylist backend.

    **This file is the integration boundary between the iOS developer and the backend/AI
    developer. Changes require both to acknowledge before merge** (PRD §16.5, CC-06).

    ## Design rules this contract encodes

    1. **The backend keeps no user data.** The device is the source of truth (PRD §11.2,
       D-13), so every request carries the data it needs and the client persists every
       result. The backend's only persistent state is a **usage ledger** — device token,
       day, task, model, cost — which exists so the daily spend cap is enforceable
       (PRD §9.6, D-20). It holds no content.
    2. **The model never sees the whole closet.** The client sends the eligible wardrobe
       subset; the backend hard-filters and scores it down to a candidate set, and the model
       selects only from that (PRD §8.1, D-03).
    3. **Outfit responses reference garments by id.** A response containing an id outside
       the candidate set is rejected by the backend before the client ever sees it
       (PRD §9.4, D-04). Such a violation is a P0 defect. **Request** payloads may still
...

    ## Privacy payload note

...

...

...

...

...

...
servers:
...
security:
  - deviceToken: []
...
paths:
...
  /v1/outfit/generate:
    post:
      tags: [outfits]
      operationId: generateOutfit
      summary: Generate a validated outfit (task T2)
      description: |
        The critical path. Runs the full pipeline: hard filter, score and shortlist, model
        selection, validation, repair, fallback (PRD §8.1).

        The client sends the wardrobe subset it can cheaply determine to be eligible
        (available, not retired, owned). The backend does the rest, so scoring can be tuned
        without an App Store round-trip (PRD §11.3).

        **The response always contains a usable outfit or an explicit gap.** It never
        contains a garment the user does not own, a draft, or an unavailable garment. Check
        `generation.fallbackLevel` to know whether the stylist model was involved at all.

        **"Try another" is stateless.** The backend keeps no record of previous outfits, so
        the client passes the garment sets already shown in this session via
        `options.excludeGarmentSets`. When no outfit differing in at least one non-locked
        slot exists, the backend returns its best outfit anyway with `noAlternativeReason`
        set, and the client shows that message instead of a silently repeated board (D-20).

        **Spend caps.** Past the soft threshold the secondary model is used. Past the hard
        cap no paid call is made and the deterministic path answers — still a 200, with
        `fallbackLevel: DETERMINISTIC`. This endpoint never returns 429.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateRequest' }
      responses:
        '200':
          description: |
            An outfit was produced. May be a model result, a repaired result, a secondary-model
            result, or a deterministic result — see `generation.fallbackLevel`. May carry
            `noAlternativeReason` when the request excluded every different outfit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerateResponse' }
        '400':
          description: |
            Malformed request, schema violation, garment not ready, or a lock/set conflict
            detected before any model call (D-32 / D-30).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              examples:
                lockConflict:
                  summary: Locked garment conflicts with anchor or availability
                  value:
                    title: Lock conflict
                    status: 400
                    code: LOCK_CONFLICT
                    detail: One or more locked assignments cannot be honoured.
                    dataPreserved: true
                    conflicts:
                      - garmentId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                        reason: Locked TOP occupies the same slot as the anchor garment.
                setConflict:
                  summary: keepTogether partner missing from wardrobe or locks
                  value:
                    title: Set conflict
                    status: 400
                    code: SET_CONFLICT
                    detail: A keepTogether set cannot be admitted whole.
                    dataPreserved: true
                    conflicts:
                      - garmentId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                        reason: Set partner 9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d is absent from wardrobe.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '422':
          description: |
            No outfit is possible from the wardrobe supplied — for example, ready coverage is
            below the minimum in PRD §7.2. This is not a failure of the model; do not retry
            without changing the input. A single unfillable slot is NOT a 422: it is a 200
            with a gap in that slot.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503': { $ref: '#/components/responses/UpstreamUnavailable' }
...
  /v1/outfit/alternatives:
    post:
      tags: [outfits]
      operationId: rankAlternatives
      summary: Rank swap alternatives for one slot against the current outfit (deterministic)
      description: |
        Owns swap re-ranking (D-33). Runs stage 1 filtering and stage 2 scoring for a single
        slot with the rest of the outfit held fixed, then applies combination rules, locks,
        and set integrity against those fixed pieces. **No model is called**; reasons are
        templated from the scoring components. Target p95 ≤ 1 s.

        The `candidateSet` returned by `/v1/outfit/generate` is valid only for the first swap
        after generation; every subsequent swap must call this endpoint, because rankings and
        reasons computed against the original outfit go stale as pieces change.

        Never returns 429: it makes no paid call.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AlternativesRequest' }
      responses:
        '200':
          description: Ranked alternatives for the requested slot. May be empty with a reason.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlternativesResponse' }
        '400':
          description: |
            Malformed request, schema violation, or a lock/set conflict against the fixed
            pieces of the current outfit (D-32 / D-30).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              examples:
                lockConflict:
                  summary: Requested swap slot is lock-fixed
                  value:
                    title: Lock conflict
                    status: 400
                    code: LOCK_CONFLICT
                    detail: The requested slot is locked and cannot be swapped.
                    dataPreserved: true
                    conflicts:
                      - garmentId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                        reason: FOOTWEAR is locked; alternatives for this slot are refused.
                setConflict:
                  summary: Alternative would split a keepTogether set
                  value:
                    title: Set conflict
                    status: 400
                    code: SET_CONFLICT
                    detail: Choosing an alternative would break a keepTogether set.
                    dataPreserved: true
                    conflicts:
                      - garmentId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                        reason: Garment is a set member; partner must move with it.
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Per-device opaque token stored only in the iOS Keychain (PRD §12.4, D-46).
        Issued by `POST /v1/auth/device` (enrollment secret); rotatable and revocable.
        No OpenRouter or provider credential ever reaches the client. Never baked into the IPA.
  responses:
...
    Unauthorized:
      description: Missing or invalid device token.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PayloadTooLarge:
      description: |
        Image payload exceeds the backend limit. The limit itself is unverified — see VF-10.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    UpstreamUnavailable:
      description: |
        The model provider is unreachable and no fallback applies. For `/v1/outfit/generate`
        this should be unreachable — the deterministic path (fallback level 3) always returns
        200. Reaching it here indicates a backend defect.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Slot:
      type: string
      description: |
        Body-coverage slots (PRD §6.1, D-22). `JACKET` (sport coat, blazer, suit jacket) and
        `OUTERWEAR` (overcoat, shell, puffer) are distinct so both can be worn on a cold day.
        There is no `FULL_BODY`: a suit is two garments linked by `setId`. `ACCESSORY` may be
        assigned up to three times; every other slot once.
      enum: [TOP, MID_LAYER, JACKET, OUTERWEAR, BOTTOM, FOOTWEAR, ACCESSORY]
...
    Boldness:
      type: string
      enum: [FAMILIAR, SLIGHT_STRETCH, BOLD]
...
    GarmentSummary:
      type: object
      additionalProperties: false
      description: |
        A garment as the backend sees it.

        **Deliberately excludes `purchasePrice`, `purchaseDate`, `priorWearEstimate`, and any
        cost-per-wear figure** (PRD §12.2). The backend cannot compute cost per wear and must
        not try; the client composes that messaging locally (D-19).

        `lastWornOn` and `wantToWearMore` are required because they drive the neglect boost
        and recency penalty (PRD §8.3) — the mechanism that gets underused garments worn
        without letting price influence selection (D-01).

        **Only READY garments may be sent** (PRD §6.1, D-23). The required fields below are
        exactly the readiness fields; a draft cannot satisfy this schema and the client must
        not invent values to make it do so. The backend rejects a request containing a
        garment with a null readiness field with 400, code `GARMENT_NOT_READY`.
      required: [id, displayName, slot, colorPrimary, pattern, surface, formality, warmth]
      properties:
        id:
          type: string
          format: uuid
          description: |
            The client's garment UUID. The backend maps these to short opaque tokens
            (`g_4f2a`) for the prompt and maps them back before responding, so the model
            never sees a UUID and validation errors are legible (PRD §9.3).
        displayName:
          type: string
          minLength: 1
          maxLength: 80
          description: |
            Never empty. User-set, or derived by the client from the readiness attributes when
            the user has not set one ("Navy sport coat"). Used verbatim in rationales (D-31).

            **Privacy (D-68 `PROPOSED`):** Sent today on image-free `generate` / `alternatives`.
            Not part of a “never sent names” guarantee for those paths until D-68 is ACCEPTED
            (or a separate redact decision lands). Still never a substitute for garment `id`
            in engine selection (design rule 3).
        slot: { $ref: '#/components/schemas/Slot' }
        category:
          type: [string, 'null']
          description: From the taxonomy in PRD App. C. Extending it is a data change, not a code change.
        colorPrimary: { $ref: '#/components/schemas/Color' }
        colorSecondary:
          oneOf:
            - { $ref: '#/components/schemas/Color' }
            - { type: 'null' }
        pattern: { $ref: '#/components/schemas/Pattern' }
        materials:
          type: array
          items: { type: string }
          maxItems: 6
        surface: { $ref: '#/components/schemas/Surface' }
        formality: { $ref: '#/components/schemas/Formality' }
        warmth: { $ref: '#/components/schemas/Warmth' }
        seasons:
          type: array
          items: { $ref: '#/components/schemas/Season' }
        fit:
          oneOf:
            - { $ref: '#/components/schemas/Fit' }
            - { type: 'null' }
        lastWornOn:
          type: [string, 'null']
          format: date
          description: Null means never worn since intake. Drives recency and neglect scoring.
        daysSinceIntake:
          type: [integer, 'null']
          description: Used for neglect scoring when `lastWornOn` is null.
        isFavorite: { type: boolean, default: false }
        wantToWearMore:
          type: boolean
          default: false
          description: The "liked but neglected" flag. Multiplies the neglect boost by 1.5.
        comfortIssue:
          type: boolean
          default: false
          description: Set by COMFORT_ISSUE feedback; lowers the score for long-day occasions.
        setId:
          type: [string, 'null']
          format: uuid
          description: |
            Links members of a coordinated set — a suit's jacket and trousers (D-22). Members
            keep their own slots. The client must send every member of a `keepTogether` set
            or none of them. The backend treats the set atomically at every stage (D-30): it
            passes stage 1 only if every member passes; it is scored as a unit (minimum of
            members) and, once shortlisted, every member enters its own slot's shortlist even
            over the per-slot cap; the model and the deterministic builder select it whole;
            validation rejects half a set. If the anchor or a locked garment is a set member
            whose partner is absent from the request, the request is rejected with 400
            `SET_CONFLICT` naming the missing partner.
        keepTogether:
          type: boolean
          default: false
          description: When true and `setId` is set, the set is used whole or not at all.
    Thumbnail:
      type: object
      additionalProperties: false
      required: [garmentId, image]
      properties:
        garmentId: { type: string, format: uuid }
        image: { $ref: '#/components/schemas/ImagePayload' }
...
    StyleProfilePayload:
      type: object
      additionalProperties: false
      description: |
        The confirmed profile as sent for generation. Null fields are omitted from the prompt
        entirely — never rendered as "unknown", which invites the model to guess (PRD §9.3).
      required: [version, summaryText, experimentationLevel]
      properties:
        version: { type: integer, minimum: 1 }
        summaryText: { type: string, maxLength: 1200 }
        age: { type: [integer, 'null'], minimum: 0, maximum: 120 }
        profession: { type: [string, 'null'], maxLength: 120 }
        workEnvironment:
          oneOf:
            - { $ref: '#/components/schemas/WorkEnvironment' }
            - { type: 'null' }
        presentationGoals:
          type: array
          items: { type: string, maxLength: 200 }
          maxItems: 8
        comfortConstraints:
          type: array
          items: { type: string, maxLength: 200 }
          maxItems: 8
        experimentationLevel: { type: integer, minimum: 1, maximum: 5 }
        activeRules:
          type: array
          items: { $ref: '#/components/schemas/PreferenceRule' }
          maxItems: 60
    ContextSnapshot:
      type: object
      additionalProperties: false
      required: [occasion, occasionFormality, temperatureBand]
      properties:
        occasion: { $ref: '#/components/schemas/Occasion' }
        occasionFormality: { $ref: '#/components/schemas/Formality' }
        temperatureBand: { $ref: '#/components/schemas/TemperatureBand' }
        precipitation: { type: boolean, default: false }
        timeOfDay:
          oneOf:
            - { $ref: '#/components/schemas/TimeOfDay' }
            - { type: 'null' }
        freeTextNote: { type: [string, 'null'], maxLength: 200 }
        capturedAt: { type: string, format: date-time }
    OutfitAssignment:
      type: object
      additionalProperties: false
      description: |
        Exactly one of `garmentId` or `gapReason` is set. A gap is named, never filled with an
        invented item or an unavailable one (PRD §7.4, D-04, D-25). `ACCESSORY` may appear up
        to three times with distinct garments; every other slot at most once (D-22).
      required: [slot]
      properties:
        slot: { $ref: '#/components/schemas/Slot' }
        garmentId: { type: [string, 'null'], format: uuid }
        isAnchor: { type: boolean, default: false }
        isLocked: { type: boolean, default: false }
        gapReason: { type: [string, 'null'], maxLength: 120 }
    Rationale:
      type: object
      additionalProperties: false
      description: |
        Length caps are the quality bar from PRD §13.6, enforced in schema rather than
        requested in prose.
      required: [summary]
      properties:
        summary: { type: string, maxLength: 200 }
        pairingNotes:
          type: array
          items: { type: string, maxLength: 140 }
          maxItems: 3
        teachingNote: { type: [string, 'null'], maxLength: 160 }
        cautions:
          type: array
          items: { type: string, maxLength: 120 }
          maxItems: 2
...
    GenerateRequest:
      type: object
      additionalProperties: false
      required: [profile, context, wardrobe]
      properties:
        profile: { $ref: '#/components/schemas/StyleProfilePayload' }
        context: { $ref: '#/components/schemas/ContextSnapshot' }
        anchorGarmentId:
          type: [string, 'null']
          format: uuid
          description: |
            When set, this garment **must** appear in the result (PRD §7.4 AC-1). Absence from
            a response is a P0 defect. Null for the daily-outfit path (M2).
        lockedAssignments:
          type: array
          description: |
            Garments the user has pinned. **Preserved exactly** — same garment, same slot — by
            the model, the deterministic builder, and validation (D-32). Capacity is the full
            outfit: six single positions plus three accessories. Each locked garment must be
            in `wardrobe`, available and ready, must not occupy the anchor's slot, and must
            not be a `keepTogether` member whose partner is absent — any of these is rejected
            with 400 `LOCK_CONFLICT` or `SET_CONFLICT` before any model call. Two locked
            garments forming a pair forbidden by an active combination rule are accepted as
            the user's explicit choice, with a caution in the rationale.
          items: { $ref: '#/components/schemas/OutfitAssignment' }
          maxItems: 9
        boldness: { $ref: '#/components/schemas/Boldness' }
        wardrobe:
          type: array
          description: |
            The eligible subset — ready (`readiness == READY`), available, not retired, owned;
            drafts must not be sent (D-23). The backend hard-filters and scores this down to
            the candidate set; the model sees only the shortlist (PRD §8.1, D-03).
          items: { $ref: '#/components/schemas/GarmentSummary' }
          minItems: 3
          maxItems: 500
        thumbnails:
          type: array
          description: |
            Optional. Whether these improve quality enough to justify cost and latency is an
            open A/B question (PRD §10.5 Q1) — not an assumption. When this array is non-empty,
            `privacyConsent` is required on the same request; the backend rejects the payload
            if `GET /v1/models` → `dataPolicy.verifiedOn` is null or consent is absent.
          items: { $ref: '#/components/schemas/Thumbnail' }
          maxItems: 13
        privacyConsent:
          description: |
            Required **only when `thumbnails` is present and non-empty**. Image-free generate
            must omit this field. Backend rejects image payloads if
            `GET /v1/models` → `dataPolicy.verifiedOn` is null **or** consent is absent when
            images are present.
          oneOf:
            - { $ref: '#/components/schemas/PrivacyConsent' }
            - { type: 'null' }
        recentOutfits:
          type: array
          description: |
            Garment id sets from suggestions in the last 14 days, for the repeat-pair penalty
            (PRD §8.3). Scoring input only; does not constrain the result.
          items:
            type: array
            items: { type: string, format: uuid }
          maxItems: 20
        options:
          type: object
          additionalProperties: false
          properties:
            candidatesPerSlot:
              type: integer
              minimum: 4
              maximum: 10
              default: 8
              description: |
                Per-slot shortlist cap. Over-cap admissions (issue #38):
                - Set partners are admitted over this cap (D-30).
                - Colour-family coverage may add up to 2 additional items (bounded by maxCap+2=12), 
                  except when boldness is FAMILIAR (coverage is skipped).
                - BOLD stretch inject adds 1 item with a distinct colour family.
                When candidatesPerSlot is explicitly set, it acts as a ceiling: BOLD boldness 
                will not increase it beyond the requested value.
            accessoryPolicy:
              type: string
              enum: [OPEN, LOCKED]
              default: OPEN
              description: |
                `OPEN` — the engine may add accessories up to three in total, keeping any locked
                ones. `LOCKED` — exactly the locked accessories, no additions (D-32). This is
                how "lock the accessory slot" is expressed, as distinct from locking one accessory.
            requireSlots:
              type: array
              items: { $ref: '#/components/schemas/Slot' }
              default: [TOP, BOTTOM, FOOTWEAR]
            excludeGarmentSets:
              type: array
              description: |
                For "try another" (D-20). Each entry is the full garment-id set of an outfit
                already shown in this dressing session. The result must differ from every
                entry in at least one non-locked slot. The backend holds no outfit history,
                so ids of previous outfits would be meaningless here — the sets themselves
                are sent. If no differing outfit exists, the backend still returns 200 with
                its best outfit and `noAlternativeReason` set.
              items:
                type: array
                items: { type: string, format: uuid }
                minItems: 2
              maxItems: 10
        requestId: { type: [string, 'null'], format: uuid }
    GenerateResponse:
      type: object
      additionalProperties: false
      description: |
        Contains no price or cost-per-wear field by design (D-19). The client composes that
        messaging from data it already holds.
      required: [outfitId, assignments, rationale, generation]
      properties:
        outfitId: { type: string, format: uuid }
        assignments:
          type: array
          items: { $ref: '#/components/schemas/OutfitAssignment' }
          minItems: 2
        rationale: { $ref: '#/components/schemas/Rationale' }
        boldnessUsed: { $ref: '#/components/schemas/Boldness' }
        noAlternativeReason:
          type: [string, 'null']
          maxLength: 200
          description: |
            Set when `options.excludeGarmentSets` ruled out every outfit that differs in a
            non-locked slot (D-20). The returned outfit may then equal one already shown; the
            client shows this reason instead of a repeated board.
        relaxationsApplied:
          type: array
          items: { $ref: '#/components/schemas/Relaxation' }
        candidateSet:
          type: object
          description: |
            The shortlist per slot, returned so the client can render the swap list without a
            second model call (PRD §7.5 AC-4). **Valid only until the first user swap that
            changes an assignment;** thereafter clients **must** call
            `POST /v1/outfit/alternatives` (D-33). Rankings and reasons computed against the
            original outfit go stale as pieces change.
          additionalProperties:
            type: array
            items:
              type: object
              additionalProperties: false
              required: [garmentId]
              properties:
                garmentId: { type: string, format: uuid }
                score: { type: [number, 'null'] }
                reason: { type: [string, 'null'], maxLength: 120 }
        generation: { $ref: '#/components/schemas/GenerationMeta' }
...
    AlternativesRequest:
      type: object
      additionalProperties: false
      required: [slot, currentAssignments, profile, context, wardrobe]
      properties:
        slot:
          $ref: '#/components/schemas/Slot'
        currentAssignments:
          type: array
          description: |
            The outfit as it stands now, including the anchor and every lock. All slots other
            than `slot` are treated as fixed pieces for scoring and rule enforcement.
          items: { $ref: '#/components/schemas/OutfitAssignment' }
          minItems: 1
          maxItems: 9
        profile: { $ref: '#/components/schemas/StyleProfilePayload' }
        context: { $ref: '#/components/schemas/ContextSnapshot' }
        wardrobe:
          type: array
          description: Eligible, ready garments, as for generation. The backend filters to `slot`.
          items: { $ref: '#/components/schemas/GarmentSummary' }
          minItems: 1
          maxItems: 500
        recentOutfits:
          type: array
          items:
            type: array
            items: { type: string, format: uuid }
          maxItems: 20
        limit: { type: integer, minimum: 1, maximum: 12, default: 8 }
        requestId: { type: [string, 'null'], format: uuid }
    AlternativesResponse:
      type: object
      additionalProperties: false
      required: [slot, alternatives]
      properties:
        slot: { $ref: '#/components/schemas/Slot' }
        alternatives:
          type: array
          description: Ranked best-first. Excludes anything violating a lock, a set, or a combination rule against the fixed pieces.
          items:
            type: object
            additionalProperties: false
            required: [garmentId, score, reason]
            properties:
              garmentId: { type: string, format: uuid }
              score: { type: number }
              reason:
                type: string
                maxLength: 120
                description: Templated from scoring components, e.g. "cooler, and keeps the brown-and-navy pairing". Not model-generated.
              setPartnerIds:
                type: array
                description: Non-empty when choosing this garment also brings its set partners into their slots.
                items: { type: string, format: uuid }
        emptyReason:
          type: [string, 'null']
          enum: [NO_ELIGIBLE, SET_BLOCKED, COMBINATION_BLOCKED, LOCK_FIXED, null]
          description: |
            Set when `alternatives` is empty. Stable codes for QA scenarios #26–28:
            `NO_ELIGIBLE` — no ready garment remains for the slot after filters;
            `SET_BLOCKED` — every candidate would split a keepTogether set;
            `COMBINATION_BLOCKED` — every candidate violates a combination rule against fixed
            pieces; `LOCK_FIXED` — the slot (or its only candidates) is locked.
...
    Problem:
      type: object
      description: |
        RFC 9457 problem detail, with product extensions (`code`, `dataPreserved`,
        `resetsAt`, conflict lists).
      additionalProperties: true
      required: [title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: |
            Stable machine-readable code. Defined codes:

            - `WARDROBE_BELOW_MINIMUM`
            - `SPEND_CAP_REACHED` (429; no-fallback paid tasks — onboarding PRD §9.3)
            - `RATE_LIMITED` (429; backoff/retry — distinct from spend cap; Workers emit today)
            - `POLICY_REJECTED` (400; consent absent/stale or `dataPolicy` unverified)
            - `BAD_REQUEST` (400; malformed / schema violation — Workers emit today)
            - `UNAUTHORIZED` (401; missing/invalid device token — Workers emit today)
            - `GARMENT_NOT_READY` (D-23)
            - `LOCK_CONFLICT` (D-32)
            - `SET_CONFLICT` (D-30)
            - `PAYLOAD_TOO_LARGE` (413)
            - `UPSTREAM_UNAVAILABLE` (503)

            `LOCK_CONFLICT` and `SET_CONFLICT` carry a `conflicts` extension listing the
            garment ids involved and a one-clause reason each. Client behaviour per
            onboarding code is in onboarding PRD §5.3.3 (CC-06).
        resetsAt:
          type: [string, 'null']
          format: date-time
          description: |
            When the client may retry or when paused copy should resume. For
            `SPEND_CAP_REACHED`, equals the ledger day end (`UsageResponse.ledgerDayEndsAt`).
            For `RATE_LIMITED`, the rate-limit window end (may be sooner than the ledger day).
        dataPreserved:
          type: boolean
          description: |
            Whether the client can safely retry without the user re-entering anything.
            Always true for a well-behaved backend — PRD §9.5 requires that no failure
            discards user input.